requirements.txt
PyQt6>=6.6.0
pytesseract>=0.3.10
PyMuPDF>=1.23.6
Pillow>=10.1.0
py-cpuinfo>=9.0.0
requests>=2.31.0
spacy>=3.7.2
llama-cpp-python>=0.2.77
# python -m spacy download en_core_web_sm